+++ src/components/deprecated/index.ts
@@ -1,2 +1,1 @@
 export { default as SAvatarMenuButton } from './SAvatarMenuButton.vue';
-export { SInputAlt, SInputDate, SInputDropdown, SInputError, SInputHelper, SInputMask, SInputRadio } from './sInputs';