+++ src/components/spartan/index.ts
@@ -39,3 +39,5 @@ export * from './SDataTable';
 export * from './SInputIncrement';
 export * from './SInputIncrementBlock';
 export * from './SSkeleton';
+export * from './SInputMask';
+export * from './SInputMaskBlock';