+++ CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Added
+- `SInputMaskBlock` component implementation.
+- `SInputMask` component implementation.
 
 ## 2.0.0-beta.56 - 2024-01-24
 ### Added